# openapi/openapi_internal_massive.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PagoPA API Debt Position",
    "description": "Progetto Gestione Posizioni Debitorie",
    "termsOfService": "https://www.pagopa.gov.it/",
    "version": "0.11.82"
  },
  "servers": [
    {
      "url": "https://api.uat.platform.pagopa.it/gpd/debt-positions-service/v1/",
      "description": "GPD Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/gpd/debt-positions-service/v1/",
      "description": "GPD Production Environment"
    },
    {
      "url": "https://api.uat.platform.pagopa.it/gpd/debt-positions-service/v3/",
      "description": "GPD Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/gpd/debt-positions-service/v3/",
      "description": "GPD Production Environment"
    },
    {
      "url": "https://api.uat.platform.pagopa.it/aca/debt-positions-service/v1/",
      "description": "ACA Test environment"
    },
    {
      "url": "https://api.platform.pagopa.it/aca/debt-positions-service/v1/",
      "description": "ACA Production environment"
    }
  ],
...
  "paths": {
...
    "/organizations/{organizationfiscalcode}/paymentoptions/{iuv}/debtposition": {
      "get": {
        "tags": [
          "Debt Positions API"
        ],
        "summary": "The organization retrieves a debt position by payment option IUV",
        "operationId": "getDebtPositionByIUV",
        "parameters": [
          {
            "name": "organizationfiscalcode",
            "in": "path",
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "required": true,
            "schema": {
              "pattern": "\\d{11}",
              "type": "string"
            }
          },
          {
            "name": "iuv",
            "in": "path",
            "description": "Payment Option IUV",
            "required": true,
            "schema": {
              "pattern": "\\b\\w{0,35}\\b",
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "examples": {
                  "forbidden": {
                    "description": "forbidden",
                    "value": {
                      "statusCode": 403,
                      "message": "You are not allowed to access this resource."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Payment Position not found.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          },
          "429": {
            "description": "Too many requests.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Service unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          },
          "200": {
            "description": "Debt Positions updated.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentPositionModelBaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Wrong or missing function key.",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ]
      },
      "parameters": [
        {
          "name": "X-Request-Id",
          "in": "header",
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "Stamp": {
        "required": [
          "hashDocument",
          "provincialResidence",
          "stampType"
        ],
        "type": "object",
        "properties": {
          "hashDocument": {
            "maxLength": 72,
            "minLength": 0,
            "type": "string",
            "description": "Document hash type is stBase64Binary72 as described in https://github.com/pagopa/pagopa-api."
          },
          "stampType": {
            "maxLength": 2,
            "minLength": 2,
            "type": "string",
            "description": "The type of the stamp"
          },
          "provincialResidence": {
            "pattern": "[A-Z]{2}",
            "type": "string",
            "description": "The provincial of the residence",
            "example": "RM"
          }
        }
      },
...
      "PaymentOptionMetadataModelResponse": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "TransferMetadataModelResponse": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "TransferModelResponse": {
        "type": "object",
        "properties": {
          "organizationFiscalCode": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "idTransfer": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "remittanceInformation": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "iban": {
            "type": "string"
          },
          "postalIban": {
            "type": "string"
          },
          "stamp": {
            "$ref": "#/components/schemas/Stamp"
          },
          "insertedDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "T_UNREPORTED",
              "T_REPORTED"
            ]
          },
          "lastUpdatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "transferMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferMetadataModelResponse"
            }
          }
        }
      },
...
      "PaymentOptionModelResponse": {
        "type": "object",
        "properties": {
          "nav": {
            "type": "string"
          },
          "iuv": {
            "type": "string"
          },
          "organizationFiscalCode": {
            "type": "string"
          },
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "isPartialPayment": {
            "type": "boolean"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "retentionDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "reportingDate": {
            "type": "string",
            "format": "date-time"
          },
          "insertedDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentMethod": {
            "type": "string"
          },
          "fee": {
            "type": "integer",
            "format": "int64"
          },
          "notificationFee": {
            "type": "integer",
            "format": "int64"
          },
          "pspCompany": {
            "type": "string"
          },
          "idReceipt": {
            "type": "string"
          },
          "idFlowReporting": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PO_UNPAID",
              "PO_PAID",
              "PO_PARTIALLY_REPORTED",
              "PO_REPORTED"
            ]
          },
          "lastUpdatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentOptionMetadata": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentOptionMetadataModelResponse"
            }
          },
          "transfer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferModelResponse"
            }
          }
        }
      },
      "PaymentPositionModelBaseResponse": {
        "type": "object",
        "properties": {
          "iupd": {
            "type": "string"
          },
          "organizationFiscalCode": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "F",
              "G"
            ]
          },
          "companyName": {
            "type": "string"
          },
          "officeName": {
            "type": "string"
          },
          "insertedDate": {
            "type": "string",
            "format": "date-time"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "validityDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "VALID",
              "INVALID",
              "EXPIRED",
              "PARTIALLY_PAID",
              "PAID",
              "REPORTED"
            ]
          },
          "lastUpdatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "paymentOption": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentOptionModelResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "The API key to access this function app.",
        "name": "Ocp-Apim-Subscription-Key",
        "in": "header"
      },
      "Authorization": {
        "type": "http",
        "description": "JWT token get after Azure Login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}